+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Show more than just one alert if more errors/warnings happen before the alert is dismissed
 * Favicon
 * Manifest for "add to homepage" on android
+* Import function for chrome/chromium bookmarks file
 
 ### Changed